requirements.txt
django>=2.2,<2.3
djangorestframework
pytest-django==3.6.0
